Refactor to use cargo metadata instead of cargo read-manifest (deprecated)
This also gives us some workspace support (you need to choose a target from workspace)
Improves #8 but only by not crashing on workspaces. We still show one target.

The current assumption of one lib max fails when running in a workspace.
Since the current code does not allow to select a specific lib target

Also there are more target kinds than bin and lib that might be interesting to parse.

We should revist the user experience here or always show the whole workspace.